Rename Android app package to com.rebeltoolbox.rebelapp
