+++ platform/android/java/app/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.godot.game"
+    package="com.rebeltoolbox.rebelapp"
     android:versionCode="1"
     android:versionName="1.0"
     android:installLocation="auto" >
